move text and number input in input-types into their own files.
To prevent this file from getting bigger and bigger as we add more inputs.
