+++ src/inputs/components/input-types/index.js
@@ -1,0 +1,7 @@
+import TextInput from "./text-input";
+import NumberInput from "./number-input";
+
+export default {
+    text: TextInput,
+    number: NumberInput
+};
