+++ src/inputs/components/input-types/text-input.js
@@ -8,7 +8,7 @@ class TextInput extends Component {
     }
 
     render() {
-        let { value, options } = this.props.input;
+        let { value } = this.props.input;
         let currentValue = value ? value : "";
 
         return (
@@ -23,19 +23,4 @@ class TextInput extends Component {
     }
 }
 
-class NumberInput extends TextInput {
-    getType() { return "number"; }
-
-    handleChange(event) {
-        let value = parseInt(event.target.value, 10);
-        if (value !== value) {
-            return false;
-        }
-        this.props.inputUpdate(value);
-    }
-}
-
-export default {
-    text: TextInput,
-    number: NumberInput
-};
+export default TextInput;
